requirements.txt
pandas~=2.2.3
plotly~=6.0.0
MetaTrader5~=5.0.4874
numpy~=1.24.3
imageio~=2.37.0
binance~=0.3
pillow~=11.1.0
setuptools~=68.2.0
pytest~=8.3.5
dask~=2024.8.0
numba~=0.58.1
scipy~=1.13.1